Set the hasher to null in the server
The test file from the core repo now has passwords and secrets in it.
To make the server echo test pass, the server under test needs to use
a null hasher. The hasher is currently not configurable in the server
so it is set to null in this commit. Eventually, the hasher (and the
cipher) will be configurable so that a real hasher (and cipher) can be
used in production, with a null hasher (and cipher) used only in the
tests.